Grouping details about op codes in compile.c
Previously there were two arrays statically initalized to hold the list
of op code names, and the op code arg counts. This patch makes a single
array of a struct which holds both the name and the arg count.

Hopefully this change will make it easier to add op codes; less counting
of values in the array to find the right spot to add the op arg count.

Signed-off-by: why the lucky stiff <why@whytheluckystiff.net>